services/agent/requirements.txt
fastapi==0.115.6
httpx==0.27.2
prometheus_client==0.21.1
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation-fastapi==0.51b0
uvicorn[standard]==0.34.0
openai>=1.0.0
openai-agents
pyjwt[crypto]==2.10.1
